refactor(acp): extract config picker shell
